requirements.txt
discord.py==1.7.3
discord-py-slash-command==3.0.3
python-dotenv==0.19.1
requests==2.26.0
beautifulsoup4==4.10.0
